requirements.txt
numpy>=1.24,<3.0
opencv-python>=4.8
pyserial>=3.5
PyYAML>=6.0
torch>=2.0
torchaudio>=2.0
torchvision>=0.15
ultralytics==8.3.40
python-statemachine>=1.0,<3.0
